+++ src/main/resources/static/js/staff/stafflist.js
@@ -11,12 +11,13 @@ $(document).ready(function(){
 });
 
 function filter(value) {
-    $("#staffList tbody tr").filter(function() {
-        let name = $(this).children().get(0).innerText.toLowerCase();
-        let legalName = $(this).children().get(1).innerText.toLowerCase();
-        let department = $(this).children().get(2).innerText.toLowerCase();
-        let visible = name.indexOf(value) > -1 || legalName.indexOf(value) > -1 || department.indexOf(value) > -1;
-        $(this).toggle(visible)
-    });
-
+    if (value.length == 0 || value.length >= 2) {
+        $("#staffList tbody tr").filter(function() {
+            let name = $(this).children().get(0).innerText.toLowerCase();
+            let legalName = $(this).children().get(1).innerText.toLowerCase();
+            let department = $(this).children().get(2).innerText.toLowerCase();
+            let visible = name.indexOf(value) > -1 || legalName.indexOf(value) > -1 || department.indexOf(value) > -1;
+            $(this).toggle(visible)
+        });
+    }
 }